+++ src/app/eth-extended/module/reorgBannerModule.tsx
@@ -1,7 +1,7 @@
 import { IModuleDef } from "plugin-api/IModuleDef";
 import { IBlockContext } from "app/shared/context/IBlockContext";
 import { blockContextType } from "app/shared/context/blockContextType";
-import { ReorgedBlocksStore } from "app/shared/data/ReorgedBlocksStore";
+import { ReorgedBlocksStore } from "app/eth-extended/data/block/ReorgedBlocksStore";
 import { IReorgBannerProps, ReorgBanner } from "app/eth-extended/module/ReorgBanner";
 
 export const reorgBannerModule: (reorgBlocksStore: ReorgedBlocksStore) =>
